FIX: specs

+++ plugins/discourse-ai/db/migrate/20260206013736_add_ai_secret_id_to_llm_and_embedding.rb
@@ -2,8 +2,8 @@
 
 class AddAiSecretIdToLlmAndEmbedding < ActiveRecord::Migration[7.0]
   def change
-    add_column :llm_models, :ai_secret_id, :integer, null: true
-    add_column :embedding_definitions, :ai_secret_id, :integer, null: true
+    add_column :llm_models, :ai_secret_id, :bigint, null: true
+    add_column :embedding_definitions, :ai_secret_id, :bigint, null: true
 
     add_index :llm_models, :ai_secret_id
     add_index :embedding_definitions, :ai_secret_id
